Enemies will be spawned only when all the enemies in the room are dead
