+++ Assets/_Project/Scriptable Objects/Weapons/Ranged/Rifles/RWS_Boss_Enemy.asset
@@ -1,0 +1,31 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3d22cffc77194ba1885f044e07aad85f, type: 3}
+  m_Name: RWS_Boss_Enemy
+  m_EditorClassIdentifier: 
+  weaponID: ERifle1
+  weaponType: 0
+  weaponName: TBD
+  damage: 1
+  range: 75
+  attackSpeed: 6.5
+  itemIcon: {fileID: 21300000, guid: d7fbd86a4b5934d3d9e7d2a74780d0ff, type: 3}
+  fireModes: 02000000
+  reloadTime: 3
+  spread: 0.02
+  recoil: 0.02
+  magazineSize: 15
+  maxBulletCount: 240
+  burstAmount: 3
+  burstDuration: 0.01
+  trailSpeed: 100
+  missDistance: 100
