+++ Assets/_Project/Scriptable Objects/Weapons/Ranged/Rifles/RWS_Boss_Enemy.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: acc01987f984a4bba8a2972d726fc48d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
