+++ Assets/_Project/Scripts/Gameplay/PCG/Room.cs
@@ -1,4 +1,5 @@
 ﻿using _Project.Scripts.Core.Backend.Ability;
+using _Project.Scripts.Core.Enemy.EnemySpawner;
 using UnityEngine;
 
 namespace _Project.Scripts.Gameplay.PCG
@@ -24,25 +25,28 @@ namespace _Project.Scripts.Gameplay.PCG
         [SerializeField] private Exit[] exits;
 
         private EnemyDeathListener _enemyDeathListener;
+        private RoomWaveController _roomWaveController;
+        
 
         /// <summary>
         /// Initializes the room by mapping local exit positions to world positions.
         /// </summary>
         private void Awake()
         {
+            _roomWaveController = GetComponent<RoomWaveController>();
             foreach (var exit in Exits)
             {
                 exit.worldPosition = transform.position + exit.localPosition;
                 Debug.Log($"Exit at {exit.localPosition} mapped to world position {exit.worldPosition}");
             }
             
-            /*if (roomType == RoomType.Exploration)
+            if (roomType == RoomType.Exploration)
             {
                 _enemyDeathListener = new EnemyDeathListener(gameObject);
-            }*/
+            }
         }
 
-        /*private void OnEnable()
+        private void OnEnable()
         {
             if(_enemyDeathListener != null)
                 _enemyDeathListener.OnAllEnemiesDead += OnAllEnemiesDead;
@@ -52,11 +56,14 @@ namespace _Project.Scripts.Gameplay.PCG
         {
             if(_enemyDeathListener != null)
                 _enemyDeathListener.OnAllEnemiesDead -= OnAllEnemiesDead;
-        }*/
+        }
         
         private void OnAllEnemiesDead()
         {
-            //AbilitySpawner.SpawnRandomAbilities(transform, size);
+            Debug.Log("All enemies dead");
+            _roomWaveController.CanSpawnEnemies();
         }
+        
+        
     }
 }
